Moved  retrievePath and retrieveType methods to HierarchicalSampleValueAnalyzerExtractionSource
